Better support 12mm labels.
This introduces a pixel width of 96 and doesn't add an offset.  In my tests,
this is maybe not _absolutely_ perfect (printing the string "T_e‾s‾t_" has the
underscores marginally closer to the tape edge than the overlines) but it's
mighty close. There might be a pixel in it?